TA3-EfrainSudhir/tutorial-backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.0.1
cffi==1.16.0
cryptography==42.0.7
fastapi==0.109.0
h11==0.14.0
mysql-connector-python==8.3.0
passlib==1.7.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.26.0
